+++ src/vlrdevapi/_event/teams/__init__.py
@@ -1,8 +1,9 @@
-from vlrdevapi._event.teams.models import EventStageTeams, EventTeams
+from vlrdevapi._event.teams.models import EventStageTeams, EventTeams, TeamPlayer
 from vlrdevapi._event.teams.namespace import EventTeamsNamespace
 
 __all__ = [
     "EventStageTeams",
     "EventTeams",
+    "TeamPlayer",
     "EventTeamsNamespace",
 ]